[feat] refactor user service to auth service
